Create bash dir and randomize poll_rho local param
Fixes tests to check for proper poll_rho results (proper factors) and
allows the 'faster' randomized `x` in the algorithm. Makes a bash
directory for storing cool shell scripts (of the bash flavor, of
course.) Adds a stub for the naive primality test using the identity
'6k+/-1' for primes greater than 3.